+++ unstructured/partition/email.py
@@ -139,6 +139,7 @@ def partition_email(
     file: Optional[IO] = None,
     text: Optional[str] = None,
     content_source: str = "text/html",
+    encoding: Optional[str] = None,
     include_headers: bool = False,
 ) -> List[Element]:
     """Partitions an .eml documents into its constituent elements.
@@ -153,7 +154,12 @@ def partition_email(
     content_source
         default: "text/html"
         other: "text/plain"
+    encoding
+        The encoding method used to decode the text input. If None, utf-8 will be used.
     """
+    if not encoding:
+        encoding = "utf-8"
+
     if content_source not in VALID_CONTENT_SOURCES:
         raise ValueError(
             f"{content_source} is not a valid value for content_source. "
@@ -170,7 +176,7 @@ def partition_email(
     elif file is not None and not filename and not text:
         file_content = file.read()
         if isinstance(file_content, bytes):
-            file_text = file_content.decode("utf-8")
+            file_text = file_content.decode(encoding)
         else:
             file_text = file_content