+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `--wikipedia-auto-suggest` argument to the ingest CLI to disable automatic redirection
   to pages with similar names.
+* Add optional `encoding` argument to the `partition_(text/email/html)` functions.
 
 ### Fixes